refactor(clawdstrike-policy-event): extract kinds/graph_slice from edr/receipt/mod.rs
