+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/kinds/mod.rs
@@ -1,0 +1,3 @@
+mod graph_slice;
+
+pub(crate) use graph_slice::*;
