+++ crates/libs/clawdstrike-policy-event/src/edr/receipt/mod.rs
@@ -2,6 +2,7 @@ mod common;
 pub mod evidence;
 pub mod families;
 pub mod inputs;
+mod kinds;
 mod serialize;
 
 pub use evidence::*;
@@ -9,6 +10,7 @@ pub use families::*;
 pub use inputs::*;
 
 pub(crate) use common::*;
+pub(crate) use kinds::*;
 
 use std::collections::{BTreeMap, BTreeSet};
 
@@ -18,7 +20,7 @@ use hush_core::{canonicalize_json, sha256, Hash};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use super::{
+pub(crate) use super::{
     action::EndpointDecisionAction,
     actor::{
         EndpointClockState, EndpointDecisionActor, EndpointPolicySnapshot, EndpointReceiptSigner,
@@ -42,7 +44,7 @@ use super::{
         EndpointPolicySimulationReport,
     },
 };
-use super::{
+pub(crate) use super::{
     endpoint_decision_actor_content_hash, endpoint_observation_content_hash,
     endpoint_policy_delta_id, endpoint_policy_event_impact_id, endpoint_policy_event_replay_id,
     endpoint_sensor_state_content_hash, event_target_field, evidence_hash_for_value, finding,
@@ -4164,65 +4166,6 @@ fn require_evidence_bundle_manifest_evidence(
         "edgeCount",
         edge_count.to_string(),
         "evidence bundle edge count evidence",
-    )
-}
-
-fn require_graph_slice_content_hash_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    content_hash: Option<&str>,
-) -> Result<()> {
-    let content_hash =
-        content_hash.ok_or_else(|| anyhow!("graph slice content hash is required"))?;
-    require_evidence_value_hash(
-        evidence,
-        "contentHash",
-        content_hash,
-        "graph slice content hash evidence",
-    )
-}
-
-fn require_graph_slice_evidence(
-    evidence: &[EndpointReceiptEvidence],
-    signed_graph_slice_id: Option<&str>,
-    graph_slice_id: Option<&str>,
-    root_node_id: Option<&str>,
-    node_count: usize,
-    edge_count: usize,
-) -> Result<()> {
-    let signed_graph_slice_id =
-        signed_graph_slice_id.ok_or_else(|| anyhow!("graph slice signed id is required"))?;
-    let graph_slice_id = graph_slice_id.ok_or_else(|| anyhow!("graph slice id is required"))?;
-    let root_node_id =
-        root_node_id.ok_or_else(|| anyhow!("graph slice root node id is required"))?;
-    if signed_graph_slice_id != graph_slice_id {
-        return Err(anyhow!(
-            "graph slice signed id must match graph reference id"
-        ));
-    }
-    require_evidence_value_hash(
-        evidence,
-        "graphSliceId",
-        graph_slice_id,
-        "graph slice id evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "rootNodeId",
-        root_node_id,
-        "graph slice root node evidence",
-    )?;
-    require_nonempty_hashed_evidence(evidence, "sliceKind", "graph slice kind evidence")?;
-    require_evidence_value_hash(
-        evidence,
-        "nodeCount",
-        node_count.to_string(),
-        "graph slice node count evidence",
-    )?;
-    require_evidence_value_hash(
-        evidence,
-        "edgeCount",
-        edge_count.to_string(),
-        "graph slice edge count evidence",
     )
 }
